Did refactoring of frontend
