+++ components/basic/Slider.js
