+++ pages/forgotpass.page.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import axios from "axios";
-import Header from "../components/Header";
+import Header from "../components/layout/Header";
 import Head from "next/head";
 import { useRouter } from "next/router";
