+++ pages/mentorRegister.page.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
+import Header from "../components/layout/Header";
+import Footer from "../components/layout/Footer";
 import MentorForm from "../components/MentorRegistration/MentorFormRegistration";
 import Head from "next/head";
